+++ example/arpeggios.f90
@@ -1,7 +1,7 @@
 ! Forsynth: a multitracks stereo sound synthesis project
 ! License GPL-3.0-or-later
 ! Vincent Magnin, 2024-05-30
-! Last modifications: 2024-05-30
+! Last modifications: 2024-05-31
 
 ! Arpeggios played in various ways using the circles of fifths
 ! https://en.wikipedia.org/wiki/Arpeggio
@@ -111,4 +111,5 @@ program arpeggios
     call demo%mix_tracks(levels=[1.2_wp, 1._wp], pan=[-0.5_wp, +0.5_wp])
     call demo%close_WAV_file()
 
+    print *,"You can now play the file ", demo%get_name()
 end program arpeggios